requirements.txt
certifi==2020.12.5
chardet==4.0.0
et-xmlfile==1.0.1
idna==2.10
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
Pillow==8.2.0
pyTelegramBotAPI==3.7.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
xlrd==2.0.1
